Fix getFriendlyName to work with groups
This function relied on several things to be true, namely that it would
be provided a device IEEE address, and it would only perform a lookup
in the cached devices. I have made this function capable of taking a
device's IEEE address or a group's ID, and it will return whichever
isn't undefined.

This commit also runs pre-commit hooks over staged files, and updates
the pre-commt hooks.